+++ src/natural-number/to-list.ts
@@ -1,0 +1,16 @@
+import { Type, Number, DigitList, Kind } from "..";
+
+export type _$toList<
+  S extends Number.Number,
+  O extends string[] = []
+> = Number._$toString<S> extends `${infer Head}${infer Tail}`
+  ? _$toList<Tail, [...O, Head]>
+  : O extends DigitList.DigitList
+  ? O
+  : ["0"];
+
+export declare abstract class ToList extends Kind.Kind {
+  abstract f: (
+    x: Type._$cast<this[Kind._], Number.Number>
+  ) => Number._$isNatural<typeof x> extends true ? _$toList<typeof x> : never;
+}