+++ src/number/from-string.spec.ts
@@ -1,0 +1,38 @@
+import { $, Test, Number } from "..";
+
+export type FromString_Spec = [
+  /**
+   * Can parse zero.
+   */
+  Test.Expect<$<Number.FromString, "0">, 0>,
+
+  /**
+   * Can parse a valid number.
+   */
+  Test.Expect<$<Number.FromString, "42">, 42>,
+
+  /**
+   * Can parse a valid negative number.
+   */
+  Test.Expect<$<Number.FromString, "-42">, -42>,
+
+  /**
+   * Can parse a valid floating point number.
+   */
+  Test.Expect<$<Number.FromString, "42.42">, 42.42>,
+
+  /**
+   * Can parse a valid negative floating point number.
+   */
+  Test.Expect<$<Number.FromString, "-42.42">, -42.42>,
+
+  /**
+   * Can parse a large number.
+   */
+  Test.Expect<$<Number.FromString, "9007199254740991">, 9007199254740991>,
+
+  /**
+   * The 'string' type results in never.
+   */
+  Test.Expect<$<Number.FromString, string>, never>
+];