+++ src/string/is-template.ts
@@ -1,9 +1,8 @@
-import { Type, Conditional, Kind, List } from "..";
-import { _$toList } from "./to-list";
+import { Type, Conditional, Kind, List, String } from "..";
 
 export type _$isTemplate<S extends string> = string extends S
   ? false
-  : List._$some<Conditional.Equals<string>, _$toList<S>>;
+  : List._$some<Conditional.Equals<string>, String._$toList<S>>;
 
 export abstract class IsTemplate extends Kind.Kind {
   abstract f: (x: Type._$cast<this[Kind._], string>) => _$isTemplate<typeof x>;