+++ src/test/expect.ts
@@ -1,4 +1,4 @@
-import { Conditional, Test } from "..";
+import { Conditional, Type } from "..";
 
 export abstract class _ {
   abstract readonly _: symbol;
@@ -7,8 +7,8 @@ export abstract class _ {
 export type Expect<
   X extends Conditional._$equals<X, V> extends true ? V : V & _,
   V = true
-> = Test._$isNever<V> extends true
+> = Type._$isNever<V> extends true
   ? X
-  : Test._$isNever<X> extends true
+  : Type._$isNever<X> extends true
   ? Expect<X, V>
   : X;