+++ src/test/index.ts
@@ -1,3 +1,2 @@
 export * from "./expect-not";
 export * from "./expect";
-export * from "./is-never";