+++ src/test/is-never.ts
@@ -1,3 +1,0 @@
-import { Conditional } from "..";
-
-export type _$isNever<X> = Conditional._$equals<X, never>;