+++ src/type/index.ts
@@ -1,3 +1,4 @@
 export * from "./cast";
 export * from "./display";
+export * from "./is-never";
 export * from "./value-of";