+++ src/union/index.ts
@@ -1,2 +1,3 @@
+export * from "./length";
 export * from "./to-intersection";
 export * from "./to-list";